pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["KAnanev <ananevdeveloper@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.14.3"
python-dotenv = "^0.18.0"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
